Change: move util and multierror to utils package
